Refactor course service to support flat course attribute model
- Removed extra service interface and update imports
- Add new methods and logic to overwrite and delete
