+++ src/main/java/com/unityTest/courseManagement/utils/Utils.java
@@ -1,5 +1,7 @@
 package com.unityTest.courseManagement.utils;
 
+import com.unityTest.courseManagement.models.CourseAttributeName;
+
 import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.IntStream;
@@ -15,5 +17,15 @@ public class Utils {
                 .collect(HashMap::new,
                         (m, i) -> m.put(keyType.cast(entries[i]), valueType.cast(entries[i + 1])),
                         Map::putAll);
+    }
+
+    public static <E extends Enum<E>> E parseToEnum(String string, Class<E> clazz) {
+        // Convert string to enum value
+        E value = null;
+        if(string != null) {
+            try { value = Enum.valueOf(clazz, string); }
+            catch (IllegalArgumentException e) { throw new IllegalArgumentException(String.format("Not one of accepted value for %s", clazz.getCanonicalName())); }
+        }
+        return value;
     }
 }
